Delete pull-requests-progress.yml
Unfortunately, there is no safe way to make secrets available and also no option to run this workflow only from this repo. As a result when a PR from a fork comes in, the action will fail. GitHub is not planning to fix this. 👎 [Discussion](https://github.community/t/make-secrets-available-to-builds-of-forks/16166/15)